+++ src/components/DocsStatusBar.js
@@ -2,8 +2,8 @@ const vscode = require('vscode');
 
 class DocsStatusBarItem {
 	constructor() {
-		this._statusBarItem = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Right, 20);
-		this._statusBarItem.text = "$(link-external) Aqueduct Docs";
+		this._statusBarItem = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Right, 50);
+		this._statusBarItem.text = "Aqueduct Docs";
 		this._statusBarItem.tooltip = "Opens the Aquedct Documentation";
 		this._statusBarItem.command = "extension.aquStatusDocs";
         this._statusBarItem.show();